+++ src/main/java/com/wise/dental_clinic/services/DentistService.java
@@ -3,10 +3,15 @@ package com.wise.dental_clinic.services;
 import com.wise.dental_clinic.dto.DentistDTO;
 import com.wise.dental_clinic.entities.Dentist;
 import com.wise.dental_clinic.repositories.DentistRepository;
+import com.wise.dental_clinic.services.exceptions.DatabaseException;
+import com.wise.dental_clinic.services.exceptions.ResourceNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -27,7 +32,7 @@ public class DentistService {
     @Transactional(readOnly = true)
     public DentistDTO findById(Long id) {
         Optional<Dentist> result = repository.findById(id);
-        Dentist dentist = result.orElseThrow();
+        Dentist dentist = result.orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
         return new DentistDTO(dentist);
     }
 
@@ -40,15 +45,26 @@ public class DentistService {
 
     @Transactional
     public DentistDTO update(DentistDTO dto, Long id) {
-        Optional<Dentist> result = repository.findById(id);
-        Dentist entity = result.orElseThrow();
-        entityToDto(entity, dto);
-        return new DentistDTO(entity);
+        try {
+            Optional<Dentist> result = repository.findById(id);
+            Dentist entity = result.orElseThrow();
+            entityToDto(entity, dto);
+            return new DentistDTO(entity);
+        } catch (NoSuchElementException e)  {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void entityToDto(Dentist entity, DentistDTO dto) {
